ci: enforce line limits across checked files
